app/requirements.txt
FastAPI
uvicorn[standard]
pydantic
sqlalchemy
pymysql
cryptography
a2wsgi
python-dotenv
waitress